feat(types): hold the Foundry surface the base factories stand on
ApplicationV2Members, DocumentSheetV2Members and VttforgeClass move from
core into @vttforge/types. Core depends on it and re-exports the names, so
consumers importing from core see no change. The types package stops being
a stub.
