+++ apps/web/index.html
@@ -280,8 +280,8 @@
               </tr>
               <tr>
                 <td class="pkg-name">@vttforge/types</td>
-                <td class="pkg-desc">Shared types. A stub today: the Foundry surface the bases declare lives in core until it is worth a package</td>
-                <td class="pkg-status s-wip"><span class="blip"></span>stub</td>
+                <td class="pkg-desc">The Foundry surface the bases stand on: ApplicationV2Members, DocumentSheetV2Members, VttforgeClass. Core re-exports them</td>
+                <td class="pkg-status s-ship"><span class="blip"></span>on npm</td>
               </tr>
             </tbody>
           </table>
