+++ packages/types/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@vttforge/types",
   "version": "0.1.3",
-  "description": "Shared TypeScript types for VTTForge. A stub today; the Foundry surface the bases declare lives in @vttforge/core.",
+  "description": "The TypeScript surface shared across VTTForge packages: the Foundry members the sheet and application bases stand on.",
   "type": "module",
   "license": "MIT",
   "homepage": "https://github.com/vttforge/vttforge#readme",
